+++ R/get_parameters.R
@@ -425,6 +425,23 @@ get_parameters.multinom <- function(x, ...) {
 get_parameters.brmultinom <- get_parameters.multinom
 
 
+#' @export
+get_parameters.mblogit <- function(x, ...) {
+  params <- stats::coef(x)
+
+  out <- data.frame(
+    Parameter = gsub("(.*)~(.*)", "\\2", names(params)),
+    Estimate = unname(params),
+    Response = gsub("(.*)~(.*)", "\\1", names(params)),
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  text_remove_backticks(out)
+}
+
+#' @export
+get_parameters.mclogit <- get_parameters.mblogit
 
 
 #' @export