+++ tests/testthat/test-mlogit.R
@@ -125,3 +125,20 @@ if (requiet("testthat") &&
     expect_identical(find_statistic(m2), "z-statistic")
   })
 }
+
+
+test_that("mblogit and mclogit is not linear", {
+  requiet("mclogit")
+
+  mod <- mblogit(factor(gear) ~ mpg + hp, data = mtcars, trace = FALSE)
+  expect_false(model_info(mod)$is_linear)
+  expect_true(model_info(mod)$is_logit)
+  expect_true(is_model(mod))
+  expect_true(is_model_supported(mod))
+
+  mod <- mclogit(factor(gear) ~ mpg + hp, data = mtcars, trace = FALSE)
+  expect_false(model_info(mod)$is_linear)
+  expect_true(model_info(mod)$is_logit)
+  expect_true(is_model(mod))
+  expect_true(is_model_supported(mod))
+})